Drop test databases at the end of the test.

+++ tests/base.py
@@ -47,6 +47,7 @@ remote = cherrypy.lib.httputil.Host('127.0.0.1', 30001)
 mockSmtp = mock_smtp.MockSmtpReceiver()
 mockS3Server = None
 enabledPlugins = []
+usedDBs = {}
 
 
 def startServer(mock=True, mockS3=False):
@@ -54,6 +55,10 @@ def startServer(mock=True, mockS3=False):
     Test cases that communicate with the server should call this
     function in their setUpModule() function.
     """
+    # If the server starts, a database will exist and we can remove it later
+    dbName = cherrypy.config['database']['uri'].split('/')[-1]
+    usedDBs[dbName] = True
+
     server = setupServer(test=True, plugins=enabledPlugins)
 
     if mock:
@@ -86,6 +91,18 @@ def stopServer():
     """
     cherrypy.engine.exit()
     mockSmtp.stop()
+    dropAllTestDatabases()
+
+
+def dropAllTestDatabases():
+    """
+    Unless otherwise requested, drop all test databases.
+    """
+    if 'keepdb' not in os.environ.get('EXTRADEBUG', '').split():
+        db_connection = getDbConnection()
+        for dbName in usedDBs:
+            db_connection.drop_database(dbName)
+        usedDBs.clear()
 
 
 def dropTestDatabase(dropModels=True):
@@ -99,8 +116,11 @@ def dropTestDatabase(dropModels=True):
 
     if 'girder_test_' not in dbName:
         raise Exception('Expected a testing database name, but got %s' % dbName)
-    db_connection.drop_database(dbName)
-
+    if dbName in db_connection.database_names():
+        if dbName not in usedDBs and 'newdb' in os.environ.get('EXTRADEBUG', '').split():
+            raise Exception('Warning: database %s already exists' % dbName)
+        db_connection.drop_database(dbName)
+    usedDBs[dbName] = True
     if dropModels:
         model_importer.reinitializeAll()
 
@@ -111,7 +131,11 @@ def dropGridFSDatabase(dbName):
     :param dbName: the name of the database to drop.
     """
     db_connection = getDbConnection()
-    db_connection.drop_database(dbName)
+    if dbName in db_connection.database_names():
+        if dbName not in usedDBs and 'newdb' in os.environ.get('EXTRADEBUG', '').split():
+            raise Exception('Warning: database %s already exists' % dbName)
+        db_connection.drop_database(dbName)
+    usedDBs[dbName] = True
 
 
 def dropFsAssetstore(path):
@@ -635,3 +659,7 @@ def _sigintHandler(*args):
 
 
 signal.signal(signal.SIGINT, _sigintHandler)
+# If we insist on test databases not existing when we start, make sure we
+# check right away.
+if 'newdb' in os.environ.get('EXTRADEBUG', '').split():
+    dropTestDatabase(False)
